AER-6891 - Create Set Indexes with sindex-admin role.
